Add language selection to home
